Add nanopb-c and protoc prebuilt.
Nanopb is required to serialize the data written to secure storage,
Nanopb depends on protoc. Because trusty uses a different build system
then Android, it is hard to build protoc in trusty. We uses a prebuilt
version from platform/tools/.
Nanopb version is 0.2.8, need to upgrade it to 0.3.9.1. Protoc version
is 2.5, planning to check in a version 3.0 prebuilt.

Test: None
Bug: b/80498777
Change-Id: I6aeed4c15445ddd3beb0f7d9e1f882033b65aece