Add libc Rust crate
While there is already a version of libc available with the Rust prebuilt toolchains, this version of libc does not include any of the patches we have applied to the AOSP copy of libc. The AOSP version includes patches for adding Trusty support, and so must be pulled in separately.

Bug: 193278128
Change-Id: I27d4e968c8ac6dafed22572a0d88ba2029915bf1